pyproject.toml
[tool.poetry]
package-mode = false
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["Nikolay Korkushko <kakashonochka12@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
amqp = "5.3.1"
asgiref = "3.9.1"
billiard = "4.2.1"
certifi = "2025.8.3"
charset-normalizer = "3.4.3"
click = "8.2.1"
click-didyoumean = "0.3.1"
click-plugins = "1.1.1.2"
click-repl = "0.3.0"
colorama = "0.4.6"
cron-descriptor = "1.4.5"
currencyconverter = "0.18.9"
django = "5.2.4"
django-celery-beat = "2.8.1"
django-filter = "25.1"
django-timezone-field = "7.1"
djangorestframework = "3.16.0"
djangorestframework-simplejwt = "5.5.1"
dnspython = "2.7.0"
drf-yasg = "1.21.10"
eventlet = "0.40.2"
forex-python = "1.9.2"
greenlet = "3.2.4"
idna = "3.10"
inflection = "0.5.1"
kombu = "5.5.4"
pillow = "11.3.0"
prompt-toolkit = "3.0.51"
psycopg2 = "2.9.10"
pyjwt = "2.10.1"
python-crontab = "3.3.0"
python-dateutil = "2.9.0"
python-dotenv = "1.1.1"
pytz = "2025.2"
pyyaml = "6.0.2"
redis = "6.4.0"
requests = "2.32.5"
simplejson = "3.20.1"
six = "1.17.0"
sqlparse = "0.5.3"
stripe = "12.4.0"
typing-extensions = "4.14.1"
tzdata = "2025.2"
uritemplate = "4.2.0"
urllib3 = "2.5.0"
vine = "5.1.0"
wcwidth = "0.2.13"
celery = "5.5.3"
django-redis = "^6.0.0"
gunicorn = "^23.0.0"


[tool.poetry.group.dev.dependencies]
black = "25.1.0"
flake8 = "7.3.0"
mypy = "1.17.1"
isort = "6.0.1"
coverage = "7.10.3"
mccabe = "0.7.0"
pycodestyle = "2.14.0"
pyflakes = "3.4.0"
mypy-extensions = "1.1.0"
packaging = "25.0"
pathspec = "0.12.1"
platformdirs = "4.3.8"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"